Add files via upload
Added a mckenzie.garnet_peridotite lithology for the parameterisation of McKenzie and Bickle (1988) for comparison of pyMelt results with results calculated using this parameterisation. Will require checking before being committed as I did most of the differentiation without being able to check if it works. Additionally, requires the scipy package to solve TSolidus as a function of pressure, in particular scipy.optimize.fsolve.